+++ datastore/dbstore/member.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"memberserver/api/models"
 	"strings"
-	"time"
 
 	"github.com/jackc/pgx/v4"
 	"github.com/jackc/pgx/v4/pgxpool"
@@ -342,64 +341,4 @@ func (db *DatabaseStore) updateSubscriptionID(memberID string, newMember models.
 	}
 
 	return nil
-}
-
-func (db *DatabaseStore) UpdateMemberCounts() {
-	dbPool, err := pgxpool.Connect(db.ctx, db.connectionString)
-	if err != nil {
-		log.Printf("got error: %v\n", err)
-	}
-	defer dbPool.Close()
-
-	err = dbPool.QueryRow(context.Background(), memberDbMethod.updateMemberCounts()).Scan()
-	if err != nil {
-		log.Errorf("conn.Query failed: %v", err)
-	}
-}
-
-func (db *DatabaseStore) GetMemberCounts() ([]models.MemberCount, error) {
-	var memberCounts []models.MemberCount
-
-	dbPool, err := pgxpool.Connect(db.ctx, db.connectionString)
-	if err != nil {
-		log.Printf("got error: %v\n", err)
-	}
-	defer dbPool.Close()
-
-	rows, err := dbPool.Query(db.ctx, memberDbMethod.getMemberCounts())
-	if err != nil {
-		log.Errorf("error getting member counts: %v", err)
-		return memberCounts, err
-	}
-
-	defer rows.Close()
-
-	for rows.Next() {
-		var m models.MemberCount
-		err = rows.Scan(&m.Month, &m.Classic, &m.Standard, &m.Premium, &m.Credited)
-		if err != nil {
-			log.Errorf("error scanning row: %s", err)
-		}
-
-		memberCounts = append(memberCounts, m)
-	}
-
-	return memberCounts, nil
-}
-
-func (db *DatabaseStore) GetMemberCountByMonth(month time.Time) (models.MemberCount, error) {
-	var memberCount models.MemberCount
-
-	dbPool, err := pgxpool.Connect(db.ctx, db.connectionString)
-	if err != nil {
-		log.Printf("got error: %v\n", err)
-	}
-	defer dbPool.Close()
-
-	err = dbPool.QueryRow(context.Background(), memberDbMethod.getMemberCountByMonth(), month).Scan(&memberCount.Classic, &memberCount.Standard, &memberCount.Premium, &memberCount.Credited)
-	if err != nil {
-		log.Errorf("conn.Query failed: %v", err)
-	}
-
-	return memberCount, nil
 }
